chore(refactor): remove login logic from app component
