+++ src/components/project/ProjectDashboard.tsx
@@ -139,8 +139,6 @@ export const ProjectDashboard: React.SFC<ParentProps> = ({
     return new LatLng(0, 0)
   }
 
-  console.log(claims)
-
   return (
     <LayoutWrapper>
       <Container className="row">
