+++ src/index.tsx
@@ -3,7 +3,7 @@ import * as ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 import { BrowserRouter as Router } from 'react-router-dom'
 import { createPublicSiteStore } from './common/redux/store'
-import { AppConnected } from './App'
+import { AppConnected } from './modules/App/App'
 
 const store = createPublicSiteStore() as any
