+++ src/modules/account/account.actions.ts
@@ -6,14 +6,15 @@ import {
   GetOrdersAction,
   UserInfo,
 } from './types'
+import { RootState } from '../../common/redux/types'
 import { Dispatch } from 'redux'
 import Axios from 'axios'
 
-export const login = (userInfo: UserInfo, error: string): LoginAction => ({
+export const login = (userInfo: UserInfo, address: string): LoginAction => ({
   type: AccountActions.Login,
   payload: {
     userInfo,
-    loginError: { error },
+    address,
   },
 })
 
@@ -82,4 +83,44 @@ export const getOrders = (address: string) => (
       }),
     ),
   })
+}
+
+export const updateLoginStatus = () => (
+  dispatch: Dispatch,
+  getState: () => RootState,
+): void => {
+  const {
+    account: { userInfo },
+    keySafe: { keysafe },
+    ixo: { ixo },
+  } = getState()
+
+  keysafe.getInfo((error, response) => {
+    if (response) {
+      const newUserInfo = { ...response, loggedInKeysafe: true }
+
+      ixo.user.getDidDoc(newUserInfo.didDoc.did).then((didResponse: any) => {
+        if (didResponse.error) {
+          newUserInfo.ledgered = false
+          newUserInfo.hasKYC = false
+        } else {
+          newUserInfo.ledgered = true
+          newUserInfo.hasKYC = didResponse.credentials.length > 0
+        }
+
+        if (JSON.stringify(userInfo) !== JSON.stringify(newUserInfo)) {
+          Axios.get(
+            `${process.env.REACT_APP_GAIA_URL}/didToAddr/${newUserInfo.didDoc.did}`,
+          ).then(addressResponse => {
+            const address = addressResponse.data
+            dispatch(login(newUserInfo, address))
+          })
+        }
+      })
+    } else {
+      dispatch(logout())
+    }
+  })
+
+  // return
 }
